<< avec fork

+++ srcs/here_doc.c
@@ -6,7 +6,7 @@
 /*   By: estarck <estarck@student.42mulhouse.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 17:33:56 by estarck           #+#    #+#             */
-/*   Updated: 2022/06/23 14:14:47 by estarck          ###   ########.fr       */
+/*   Updated: 2022/06/23 15:50:24 by estarck          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,7 @@ static void	write_cat(t_lst *cmd)
 	char	**argv;
 
 	argv = malloc(sizeof(char *) * 2);
-	argv[0] = malloc(sizeof(char) * 4);
-	argv[1] = malloc(sizeof(char) * 1);
-	argv[0] = "cat";
+	argv[0] = ft_strdup("cat");
 	argv[1] = NULL;
 	cmd->argv = argv;
 }
@@ -65,27 +63,40 @@ static int	find_eof(char **argv)
 
 static void	run_heredocs(t_lst *cmd)
 {
+	pid_t	pid;
 	int		fd;
 	char	*buf;
 	int		i;
 
-	buf = readline("heredoc> ");
-	fd = open(cmd->tmpfile, O_TRUNC | O_CREAT | O_WRONLY, 0777);
-	i = find_eof(cmd->argv);
-	while (1)
+	pid = fork();
+	if (pid == 0)
 	{
-		if (ft_strcmp(buf, cmd->argv[i + 1]) == 0)
-			break ;
-		write(fd, buf, ft_strlen(buf));
-		write(fd, "\n", 1);
 		buf = readline("heredoc> ");
+		fd = open(cmd->tmpfile, O_TRUNC | O_CREAT | O_WRONLY, 0777);
+		i = find_eof(cmd->argv);
+		while (42)
+		{
+			if (ft_strcmp(buf, cmd->argv[i + 1]) == 0)
+				break ;
+			write(fd, buf, ft_strlen(buf));
+			write(fd, "\n", 1);
+			buf = readline("heredoc> ");
+		}
+		free(buf);
+		close(fd);
+		cmd->pipefd[READ] = open(cmd->tmpfile, O_RDONLY, 0777);
+		unlink(cmd->tmpfile);
+		close(cmd->pipefd[WRITE]);
+		dup2(cmd->pipefd[READ], STDIN_FILENO);
+		close(cmd->pipefd[READ]);
+		return ;
 	}
-	free(buf);
-	close(fd);
-	cmd->pipefd[READ] = open(cmd->tmpfile, O_RDONLY, 0777);
-	unlink(cmd->tmpfile);
-	dup2(cmd->pipefd[READ], STDIN_FILENO);
-	close(cmd->pipefd[READ]);
+	else
+	{
+		waitpid(-1, 0, 0);
+		close(cmd->pipefd[WRITE]);
+		close(cmd->pipefd[READ]);
+	}
 }
 
 void	init_heredoc(t_lst *cmd)
